feat: add services for get unique instances and get full by unique
The first service is needed because you cannot make bulk_create if the model instances are duplicated in the list.
The second service is needed in order to return the original list with repeats after saving unique instances, but with their IDs from the database. IT specialists are needed to further save other instances that link to these.

Refs: #37